FEAT: Creation of tests for StreamFactory and in consequence much tidying and refactoring of all the things

+++ device-detection-core/src/main/java/fiftyone/mobile/detection/cache/ICache.java
@@ -12,7 +12,7 @@ package fiftyone.mobile.detection.cache;
  * @author jo
  */
 public interface ICache<K,V> {
-    int getCacheSize();
+    long getCacheSize();
 
     long getCacheMisses();
 
@@ -32,7 +32,7 @@ public interface ICache<K,V> {
 
     abstract class Base <K, V> implements ICache<K, V> {
         @Override
-        public int getCacheSize() {
+        public long getCacheSize() {
             return 0;
         }
